Workaround pio bug that dropped -DENABLE_WIFI from the g++ command
If the !python git-version.py  line is at the end of the build_flags
block, the next build flag from an env section is spliced onto the
end of the python command line instead of being passed on to g++ as
a separate flag, thus ignoring that next build flag.

The workaround is to move the !python line up in the list so it
is not at the end.