Fixed bug in setting completion
To reproduce the problem, just after starting Fluidterm,
type $<tab>.  Instead of it completing to $Firmware/Build,
it will complete to $irmware/Build without the initial F,
and if you then hit enter, it will execute just "$".

The problem was caused by a static string that was overwritten
at an inopportune time.