Automatic reporting experiment
This is a proof of concept for automatic status reports. To use it:

$Report/Interval=400    (milliseconds, $SRI is short name)

That makes the channel on which the command was received act as
if ? were sent every N milliseconds.

Deficiences:
* There is no attempt to omit unchanged fields
* Modal info like G20 is not auto-reported
* The interval is the same in run and idle states

It seems to work pretty well with WebUI.  You can un-check the
"auto-check every ..." box to turn off ? sending, then manually
send "$report/interval=200" via Commands > Send (or MDI in tablet)
and things will work.