Use intrinsic atomics
Possible fix for #472

It looks like we got bitten once again by two perennial sources of
problems - atomics and FLASH ISRs.  For compatibility with the new
framework and toolchain, we had to switch from the old "stdatomic"
to the new <atomic>, but the latter uses subroutine calls for
atomic access instead of putting the synchronizing instructions
inline.  The subroutines in question are in FLASH, not IRAM, so
if you call them from ISRs, the system can crash.

So I added the ability to use GCC atomic intrinsics as an alternative
to <atomic>.  The intrinsics compile inline code.