Network and Telnet cleanup
The big change here is in the Telnet code.  There is now a
TelnetClient class that derives from Channel.  When a Telnet
connection comes in, TelnetServer creates a TelnetClient instance
and registers it with allChannels.  This makes Telnet behave like
other channels.  It also enables multiple simultaneous Telnet
connections that play well with each other.