Use std::filesystem instead of Arduino FS classes
This is a major cleanup of the code that deals with filesystems
like SDFAT, SPIFFS, and LittleFS.  Instead of using the Arduino
filesystem classes like SD and SPIFFS, the interface is through
the C++ std::filesystem class, which is in turn implemented on
top of ESP-IDF and stdio.

Many FluidNC filesystem access functions that used to be
duplicated for different filesystem types are now common.
There is a key interface layer "FluidPath" that derives from
std::filesystem::path.  Its job is to combine a filesystem
name like "/sd" with a path, mount the appropriate filesystem,
and keep it mounted until the FluidPath object goes away,
so the filesystem stays alive while it is being used.  This
replaces the tricky SD.begin() / SD.end() bracketing that
was previously required.

ESP-IDF does not yet have the appropriate library -lstdc++fs,
so I grabbed the GNU source code for it and included it in
the subdirectory stdfs/ .  The modifications to that source
code were nearly trivial.  Nearly everything just worked,
but I did have to do some work to implement std::filesystem::space().

This patch is a step toward the long-term goal of reducing/eliminating
the dependence on the Arduino framework.