bug: Resolve broken $I command on serial console
Presently the `$I` command fails to reset the contents of the variable
`result` due to a static definition in the funciton
`WiFiConfig::info()`.  Calling the function repeatedly causes
repetitions in the output:

```
$I
[VER:3.5 FluidNC v3.5.0:]
[OPT:MPHS]
[MSG: Machine: CNC_xPRO_V5]
[MSG: Mode=STA:SSID=2e:Status=Connected:IP=192.168.132.163:MAC=98-CD-AC-B4-10-78]
ok
$I
[VER:3.5 FluidNC v3.5.0:]
[OPT:MPHS]
[MSG: Machine: CNC_xPRO_V5]
[MSG: Mode=STA:SSID=2e:Status=Connected:IP=192.168.132.163:MAC=98-CD-AC-B4-10-78Mode=STA:SSID=2e:Status=Connected:IP=192.168.132.163:MAC=98-CD-AC-B4-10-78]
ok
$I
[VER:3.5 FluidNC v3.5.0:]
[OPT:MPHS]
[MSG: Machine: CNC_xPRO_V5]
[MSG: Mode=STA:SSID=2e:Status=Connected:IP=192.168.132.163:MAC=98-CD-AC-B4-10-78Mode=STA:SSID=2e:Status=Connected:IP=192.168.132.163:MAC=98-CD-AC-B4-10-78Mode=STA:SSID=2e:Status=Connected:IP=192.168.132.163:MAC=98-CD-AC-B4-10-78]
```

Removing `static` will change the scope of `result` to a single instance
and not share the value across all instances.