format: Run `clang-format`
Running `clang-format` to ensure proper formatting.  In this case it
seems `clang-format` may not have been run on this file previously.

As the option `AllowShortFunctionsOnASingleLine` is set to the value
`Inline` and we are not assigning these values as a part of the class
definition, it is expanding the function definitions to a multi-line
syntax.