Multiple WebSocket channels
It is now possible to have multiple incoming websockets from different
hosts, without interference from WebUI websockets.  Furthermore, it
is possible to connect multiple instances of WebUI to FluidNC at the
same time, either from the same host computer or from different
computers, in any combination.

If you are connecting from a program other than WebUI, supply a URI
other than just "/".  If the URI is /, the WebUI PING / ACTIVE_ID
protocol will be used.  That form of PING is not the same as standard
WebSocket Ping packets.