Create build-only CI workflow
This CI workflow triggers a Github action to:
- Install PlatformIO at a known version (modulo minor revision)
- Runs against Windows, Linux, and MacOS
- Builds a number of PlatformIO environments

The PlatformIO environments built are initially noradio, wifi, and bt.
It is easy to include additional environments, but in the spirit of having
a CI which balances speed with catching true-positive failures, we select
only these three.

In addition, there is support for additional "_s2", "_s3" variants, but
for this initial CI pipeline, they are not built, as they greatly increase
CI time.