Fixed single-axis homing fail with cycle=0.
The fix involved adding homing/cycle=0 axes to the homing mask.
The problem was introduced via a combination of automatically
creating homing sections if they were not explicitly present,
along with the way that the default cycle value for such sections
was interpreted for the purpose of homing mask.  I eliminated the
inferred sections, ensuring that code that uses homing sections
is careful to check for nonexistent ones.  In the process, I factored
the tstep calculation in Trinamic drivers so the homing related code
is in one place.