standarized on one delay(ms) call
The code base currently has 3 calls to delay for X number of milliseconds
This change standarized on the default delay(ms) call which just
call vTaskDelay() under the covers.  It also updates the x86 support
to do the same.  The custom delay_ms(ms) in the code base has the same
implementation as delay(ms), so those calls have been updated to just
use delay(ms).