conditional compilation / inclusion of OLED support
because most boards don't have an OLED display,  the OLED module is
retired, the OLED library uses a fair amount of ROM, and it uses
the String class, provide an option to not included it.

build_src_filter is not an option because of other dependencies in
the source code on the OLED class existence.