Added command line upload for fluidterm
for example

   fluidterm -p COM3 -u files\myprog.gc -r /sd/program1.gc

upload files\myprog.gc from the host machine, writing it
to /sd/program1.gc on the FluidNC machine.

If -r (the remote name) is omitted, the uploaded file will be
placed in localfs with a name the same as the tail name of the
source file, for example /localfs/myprog.gc per above.

If the remote name ends with /, it is the remote directory in
which to place the uploaded file, with the same tail name as
the source file.  For example

fluidterm -p COM3 -u files\myprog.gc -r /sd/

... would write the file to /sd/myprog.gc

If -p (the serial port name) is omitted and there is only one
candidate port, that port will be used automatically.  If there
are multiple candidates, fluidterm will prompt you to choose one.
Obviously that is not particularly suitable for use in a script
or batch file, so it is best to supply -p in that case.